+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 ## Unreleased
 
+### Changed
+- Removed `dukt\social\services\LoginAccounts::saveLoginAccount()`. ([#26](https://github.com/dukt/social/issues/26))
+
 ### Fixed
 - Fixed a bug where registering a user with a social account that was used for a soft-deleted user was throwing an error, preventing completion of the registration process. ([#6](https://github.com/dukt/social/issues/6))